Reverted Xamarin.Forms to fix startup problem.
On launch of the Android test runner, there was a Xamarin.Forms
exception that prevented it from starting.  This seemed to be related to
the discussion in xunit/devices.xunit#37.  Switching back to the lowest
possible 1.5 version of Xamarin.Forms allowed by devices.xunit solved
that problem.  Setting all of the Android target versions to 21 resolved
a method missing exception.